began to pick which possible arrangement of the two pairs in the quad should be selected based on maximizing the difference between the masses of the pairs
added histogram of the sum of pair_12_34_ZOnly, pair_13_24_ZOnly, and pair_14_23_ZOnly. Confirmed that as expected in the Z->4 mu case, this sum is nearly always 2. Began to sort out which arrangement of the two pairs in the quad should be selected by maximizing the difference between the masses of the pairs. Dealt with the case where pair_12_34_ZOnly and pair_13_24_ZOnly are true.